64-bit patches (Fraser Speirs and Ferruccio Vitale), plus fixed a memory leak (Trevor Squires) and fixed issue with the accessoryView's resizing mask not being fully updated (Dan Wood).
git-svn-id: http://svn.cocoasourcecode.com/MGScopeBar@43 fad7f400-0e54-0410-bf1d-e368f886e4d4

+++ MGScopeBar.h
@@ -18,7 +18,7 @@
 	NSMutableDictionary *_identifiers; // map of identifiers to items.
 	NSMutableArray *_selectedItems; // all selected items in all groups; see note below.
 	float _lastWidth; // previous width of view from when we last resized.
-	int _firstCollapsedGroup; // index of first group collapsed into a popup.
+	NSInteger _firstCollapsedGroup; // index of first group collapsed into a popup.
 	float _totalGroupsWidthForPopups; // total width needed to show all groups expanded (excluding padding and accessory).
 	float _totalGroupsWidth; // total width needed to show all groups as native-width popups (excluding padding and accessory).
 	BOOL _smartResizeEnabled; // whether to do our clever collapsing/expanding of buttons when resizing (Smart Resizing).
